requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
SQLAlchemy==1.4.25
mysql-connector-python==8.0.27
Jinja2~=3.1.2
tldextract~=5.1.1
colorama~=0.4.6